followanalytics/fa-issues#4891 Improved eval booster code
- Added some fallbacks and refactored template code
